Masalar ekranında her bir masa için toplam tutarın görüntülenmesi sağlandı

+++ src/main/java/com/ergintuncer/restaurant/entity/Masa.java
@@ -21,7 +21,7 @@ public class Masa implements Serializable {
     @Column(name = "masa_kapasite", nullable = false)
     private short masaKapasite;
     @Transient
-    private float toplamFiyat;
+    private double toplamFiyat;
     @OneToMany(mappedBy = "masa")
     private List<Siparis> siparis;
